xts_tbl

This function converts data from a xts object to a tbl_df(). Note that the dataframe must be of type xts and ordered by a date column. This date column will be preserved and save as "date".

xts_tbl(xts)

Arguments

xts
A xts series that will be converted to a tbl_df().

Value

A tbl_df() with the first column the "date" column used to order the xts series by.

Examples

library(dplyr)
data(TRI)
TRI %>% tbl_xts(., cols_to_xts = "TRI", spread_by = "Country") %>% xts_tbl()

#> # A tibble: 790 x 22
#>    date         brz   chl   chn   col   cze   egt   grc   hun   ind   ino   kor
#>    <date>     <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 2000-01-14 1420.   NA   36.7 105.   91.7  352. 1215.  369.  175.  280.  192.
#>  2 2000-01-21 1382.   NA   36.1 102.   98.3  375. 1177.  379.  179.  261.  186.
#>  3 2000-01-28 1344.   NA   36.0 104.   97.1  354. 1125.  379.  187.  246.  191.
#>  4 2000-02-04 1426. 1522.  35.5 105.  104.   349. 1202.  370.  198.  239.  189.
#>  5 2000-02-11 1416. 1544.  31.8 104.  114.   359. 1257.  396.  220.  249.  188.
#>  6 2000-02-18 1372. 1443.  33.2  97.7 114.   350. 1209.  402.  243.  231.  172.
#>  7 2000-02-25 1385. 1380.  29.7  90.3 112.   322. 1195.  405.  222.  219.  168.
#>  8 2000-03-03 1439. 1456.  28.5  88.2 117.   327. 1164.  401.  217.  208.  177.
#>  9 2000-03-10 1424. 1421.  29.8  88.4 117.   327. 1111.  425.  211.  234.  174.
#> 10 2000-03-17 1383. 1419.  29.9  87.9 118.   328. 1099.  399.  189.  230.  169.
#> # ... with 780 more rows, and 10 more variables: mal <dbl>, mex <dbl>,
#> #   per <dbl>, phi <dbl>, pol <dbl>, rus <dbl>, zar <dbl>, twn <dbl>,
#> #   tha <dbl>, tur <dbl>